extract types to separate file
